Correct UI and mesh render order

+++ Sources/Kernel.cpp
@@ -9,7 +9,7 @@ float Kernel::GetValue(float distance) const
     else
     {
         float x = 1.0f - distance * distance / _r2;
-        return (315.0f * x * x * x) / (64.0f * M_PI * _r3);
+        return static_cast<float>((315.0f * x * x * x) / (64.0f * M_PI * _r3));
     }
 }
 
@@ -22,7 +22,7 @@ float Kernel::FirstDerivative(float distance) const
     else
     {
         float x = 1.0f - distance / _r1;
-        return -45.0f * x * x / (M_PI * _r4);
+        return static_cast<float>(-45.0f * x * x / (M_PI * _r4));
     }
 }
 
@@ -35,7 +35,7 @@ float Kernel::SecondDerivative(float distance) const
     else
     {
         float x = 1.0f - distance / _r1;
-        return 90.0f * x / (M_PI * _r5);
+        return static_cast<float>(90.0f * x / (M_PI * _r5));
     }
 }
